Extracted enums to separate project
