+++ Mathema.Enums/Constants/Constants.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Mathema.Enums.Constants
+{
+    public enum ConstantTypes
+    {
+        PI,
+        e,
+        EM,
+        Phi
+    }
+}
