+++ Mathema.Enums/Functions/Functions.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Mathema.Enums.Functions
+{
+    public class Functions
+    {
+        public enum FunctionTypes
+        {
+            Sin = 1,
+            Cos = 2,
+            Tan = 3,
+            Ctan = 4
+        }
+    }
+}
