+++ Mathema.Enums/Symbols/Symbols.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Mathema.Enums.Symbols
+{
+    public enum SymbolTypes
+    {
+        Undefined = 0,
+        Number = 1,
+        UnaryOperator = 2,
+        BinaryOperator = 3,
+        LeftParenthesis = 4,
+        RightParenthesis = 5,
+        Function = 6,
+        Variable = 7
+    }
+
+}
