+++ Mathema.Interfaces/ISymbol.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Mathema.Enums.Symbols;
+using System;
 using System.Collections.Generic;
 using System.Text;
 
@@ -7,5 +8,7 @@ namespace Mathema.Interfaces
     public interface ISymbol
     {
         string Value { get; set; }
+
+        SymbolTypes Type { get; set; }
     }
 }
