+++ Mathema.Interfaces/Mathema.Interfaces.csproj
@@ -4,4 +4,8 @@
     <TargetFramework>netcoreapp2.1</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Mathema.Enums\Mathema.Enums.csproj" />
+  </ItemGroup>
+
 </Project>
