+++ Mathema.Models/Constants/Constant.cs
@@ -1,4 +1,4 @@
-﻿using Mathema.Shared.Constants;
+﻿using Mathema.Enums.Constants;
 using System;
 using System.Collections.Generic;
 using System.Text;
