+++ Mathema.Models/Expressions/FunctionExpression.cs
@@ -3,6 +3,7 @@ using Mathema.Models.Functions;
 using System;
 using System.Collections.Generic;
 using System.Text;
+using static Mathema.Enums.Functions.Functions;
 
 namespace Mathema.Models.Expressions
 {
