+++ Mathema.Models/Functions/Functions.cs
@@ -1,17 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using static Mathema.Enums.Functions.Functions;
 
 namespace Mathema.Models.Functions
 {
-    public enum FunctionTypes
-    {
-        Sin = 1,
-        Cos = 2,
-        Tan = 3,
-        Ctan = 4
-    }
-
     public class Functions
     {
         public static Dictionary<string, Function> All { get; } = GetAllFunctions();
