+++ Mathema.Models/Symbols/Symbols.cs
@@ -1,21 +1,10 @@
-﻿using System;
+﻿using Mathema.Enums.Symbols;
+using System;
 using System.Collections.Generic;
 using System.Text;
 
 namespace Mathema.Models.Symbols
 {
-    public enum SymbolTypes
-    {
-        Undefined = 0,
-        Number = 1,
-        UnaryOperator = 2,
-        BinaryOperator = 3,
-        LeftParenthesis = 4,
-        RightParenthesis = 5,
-        Function = 6,
-        Variable = 7
-    }
-
     public class Symbols
     {
         private static readonly Dictionary<string, Symbol> All = GetAllSymbols();
