+++ Mathema.sln
@@ -19,7 +19,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mathema.RPNTests", "Mathema
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExpressionTests", "ExpressionTests\ExpressionTests.csproj", "{DD47D115-B176-45A8-A103-9984651CBADE}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FlatExpressionTests", "FlatExpressionTests\FlatExpressionTests.csproj", "{3C2C9AB3-FBC5-47D3-A2EC-EB0F8E279725}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FlatExpressionTests", "FlatExpressionTests\FlatExpressionTests.csproj", "{3C2C9AB3-FBC5-47D3-A2EC-EB0F8E279725}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Mathema.Enums", "Mathema.Enums\Mathema.Enums.csproj", "{446FB13E-0EEA-461A-8B68-AE1D5A6D3B15}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -59,6 +61,10 @@ Global
 		{3C2C9AB3-FBC5-47D3-A2EC-EB0F8E279725}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3C2C9AB3-FBC5-47D3-A2EC-EB0F8E279725}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3C2C9AB3-FBC5-47D3-A2EC-EB0F8E279725}.Release|Any CPU.Build.0 = Release|Any CPU
+		{446FB13E-0EEA-461A-8B68-AE1D5A6D3B15}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{446FB13E-0EEA-461A-8B68-AE1D5A6D3B15}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{446FB13E-0EEA-461A-8B68-AE1D5A6D3B15}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{446FB13E-0EEA-461A-8B68-AE1D5A6D3B15}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
